# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cold Chain Integrated Prototype API",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v1/control/{device_id}/pending": {
      "get": {
        "tags": [
          "Project compatibility"
        ],
        "summary": "Pending",
        "operationId": "pending_api_v1_control__device_id__pending_get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "x-device-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Device-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/devices": {
      "post": {
        "summary": "Register",
        "operationId": "register_api_devices_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegistration"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Devices",
        "operationId": "devices_api_devices_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/telemetry": {
      "post": {
        "summary": "Telemetry",
        "operationId": "telemetry_api_telemetry_post",
        "parameters": [
          {
            "name": "x-device-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Device-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Telemetry"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceRegistration": {
        "properties": {
          "device_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,48}$",
            "title": "Device Id"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "mode": {
            "type": "string",
            "enum": [
              "SIMULATION",
              "HARDWARE"
            ],
            "title": "Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "device_id",
          "name",
          "mode"
        ],
        "title": "DeviceRegistration"
      },
      "GPS": {
        "properties": {
          "fix": {
            "type": "boolean",
            "title": "Fix",
            "default": false
          },
          "source": {
            "type": "string",
            "enum": [
              "GPS",
              "SIMULATED",
              "NONE"
            ],
            "title": "Source",
            "default": "NONE"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "speed_kmph": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 400.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Kmph"
          },
          "satellites": {
            "type": "integer",
            "maximum": 99.0,
            "minimum": 0.0,
            "title": "Satellites",
            "default": 0
          },
          "age_s": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age S"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "GPS"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SensorHealth": {
        "properties": {
          "chamber": {
            "type": "boolean",
            "title": "Chamber"
          },
          "heatsink": {
            "type": "boolean",
            "title": "Heatsink"
          },
          "sht31": {
            "type": "boolean",
            "title": "Sht31"
          },
          "current": {
            "type": "boolean",
            "title": "Current"
          },
          "door": {
            "type": "boolean",
            "title": "Door"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "chamber",
          "heatsink",
          "sht31",
          "current",
          "door"
        ],
        "title": "SensorHealth"
      },
      "Telemetry": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "device_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,48}$",
            "title": "Device Id"
          },
          "boot_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,64}$",
            "title": "Boot Id"
          },
          "sequence": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Sequence"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "uptime_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Uptime Ms"
          },
          "mode": {
            "type": "string",
            "enum": [
              "SIMULATION",
              "HARDWARE"
            ],
            "title": "Mode"
          },
          "chamber_temp_c": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 125.0,
                "minimum": -55.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Chamber Temp C"
          },
          "heatsink_temp_c": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 125.0,
                "minimum": -55.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Heatsink Temp C"
          },
          "sht31_temp_c": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 125.0,
                "minimum": -40.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sht31 Temp C"
          },
          "humidity_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Humidity Pct"
          },
          "primary_current_a": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 20.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Current A"
          },
          "door_open": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Door Open"
          },
          "door_open_s": {
            "type": "number",
            "minimum": 0.0,
            "title": "Door Open S",
            "default": 0
          },
          "gps": {
            "$ref": "#/components/schemas/GPS"
          },
          "primary_cooling": {
            "type": "boolean",
            "title": "Primary Cooling"
          },
          "backup_cooling": {
            "type": "boolean",
            "title": "Backup Cooling"
          },
          "system_state": {
            "type": "string",
            "enum": [
              "NORMAL",
              "WARNING",
              "PRIMARY_FAULT",
              "BACKUP_ACTIVE",
              "RECOVERY",
              "CRITICAL_FAILURE",
              "REROUTING"
            ],
            "title": "System State"
          },
          "sensor_health": {
            "$ref": "#/components/schemas/SensorHealth"
          },
          "fault_injection": {
            "type": "string",
            "enum": [
              "NONE",
              "PRIMARY_FAILURE",
              "BACKUP_FAILURE"
            ],
            "title": "Fault Injection",
            "default": "NONE"
          },
          "buffered": {
            "type": "boolean",
            "title": "Buffered",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "device_id",
          "boot_id",
          "sequence",
          "timestamp",
          "uptime_ms",
          "mode",
          "chamber_temp_c",
          "heatsink_temp_c",
          "sht31_temp_c",
          "humidity_pct",
          "primary_current_a",
          "door_open",
          "gps",
          "primary_cooling",
          "backup_cooling",
          "system_state",
          "sensor_health"
        ],
        "title": "Telemetry"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}